requirements.txt
aiobotocore==1.3.0
aiofiles==0.6.0
aiohttp==3.7.3
aioitertools==0.7.1
alabaster==0.7.12
anyio==3.3.0
apipkg==1.5
appdirs==1.4.4
argcomplete==1.10.0
argon2-cffi==20.1.0
asn1crypto==1.3.0
asv==0.4.2
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
awscli==2.0.61
Babel==2.9.0
backcall==0.1.0
backports.csv==1.0.7
banal==1.0.1
Beaker==1.10.0
beautifulsoup4==4.9.1
black==20.8b1
bleach==3.3.0
blivet==3.2.2
blivet-gui==2.1.15
blosc==1.10.2
boto3==1.17.57
botocore==2.0.0.dev65
Bottleneck==1.3.2
Brlapi==0.7.0
cachetools==4.2.1
certifi==2021.5.30
cffi==1.14.0
cfgv==3.2.0
cftime==1.4.1
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.1
colour==0.1.5
coverage==5.5
cpplint==1.5.4
cryptography==3.4.7
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.23
dasbus==0.2
dask==2021.4.1
dbus-python==1.2.16
decorator==4.4.0
defusedxml==0.7.1
distlib==0.3.1
distro==1.5.0
docutils==0.15.2
ec2instanceconnectcli==1.0.2
entrypoints==0.3
et-xmlfile==1.0.1
execnet==1.8.0
fastparquet==0.5.0
feedparser==5.2.1
filelock==3.0.12
flake8==3.9.1
flake8-bugbear==21.3.2
flake8-comprehensions==3.1.0
Flask==1.1.2
fros==1.1
fsspec==2021.4.0
future==0.18.2
gcsfs==2021.4.0
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.29.0
google-auth-oauthlib==0.4.4
gpg==1.14.0
greenlet==1.0.0
Guake==3.6.3
html5lib==1.1
humanize==0.5.1
hypothesis==6.10.0
identify==2.2.4
idna==2.8
imagesize==1.2.0
iniconfig==1.1.1
iniparse==0.4
iotop==0.6
IPy==0.81
ipykernel==5.5.3
ipython==7.12.0
ipython-genutils==0.1.0
ipywidgets==7.6.3
isc==2.0
isort==5.8.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.11.3
jmespath==0.9.4
joblib==1.0.1
json5==0.9.6
jsonlines==2.0.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.10.1
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.1
jupyterlab-widgets==1.0.0
kaggle==1.5.12
kiwisolver==1.3.1
langtable==0.0.52
libcomps==0.1.15
libvirt-python==6.1.0
llvmlite==0.36.0
locket==0.2.1
lxml==4.4.1
mac-vendor-lookup==0.1.11
Mako==1.1.1.dev0
manimlib==0.2.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
meld==3.20.2
meson==0.55.1
mistune==0.8.4
more-itertools==8.7.0
moto==2.0.5
multidict==5.1.0
musicbrainzngs==0.5
mutagen==1.43.0
mypy==0.812
mypy-extensions==0.4.3
natsort==7.1.1
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.0.8
nbsphinx==0.8.3
nest-asyncio==1.5.1
nftables==0.1
nodeenv==1.6.0
normality==2.1.1
notebook==6.3.0
ntplib==0.3.3
numba==0.53.1
numexpr==2.7.3
numpy==1.20.2
numpydoc==1.1.0
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
opencv-python==4.5.3.56
openpyxl==3.0.7
ordered-set==3.1
packaging==20.4
pandas==1.3.0.dev0+1413.ga9c7898cc7
pandas-dev-flaker==0.2.0
pandoc==1.0.2
pandocfilters==1.4.3
parso==0.5.2
partd==1.2.0
Paste==3.2.4
pathspec==0.8.1
patsy==0.5.1
pbr==5.4.3
pexpect==4.7.0
pgcsv==0.4
pickleshare==0.7.5
pid==2.2.3
Pillow==7.0.0
pipx==0.15.5.1
pluggy==0.13.1
ply==3.11
pre-commit==2.12.1
productmd==1.27
progressbar==2.5
prometheus-client==0.10.1
prompt-toolkit==2.0.10
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pwquality==1.4.2
py==1.10.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.7.0
pycoingecko==2.2.0
pycparser==2.19
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.43.0.5
pydata-sphinx-theme==0.6.1
pydbus==0.6.0
pydub==0.25.1
pyenchant==2.0.0
pyflakes==2.3.1
Pygments==2.4.2
PyGObject==3.36.1
pyinotify==0.9.6
pykickstart==3.24
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyparted==3.11.4
PyQt5==5.14.2
PyQt5-sip==4.19.21
pyreadstat==1.1.0
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.2.3
pytest-asyncio==0.15.1
pytest-cov==2.11.1
pytest-forked==1.3.0
pytest-instafail==0.4.2
pytest-xdist==2.2.1
python-augeas==0.5.0
python-dateutil==2.8.0
python-meh==0.48
python-slugify==5.0.2
python-snappy==0.6.0
pytz==2020.1
pyudev==0.22.0
pyupgrade==2.13.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.0.3
quodlibet==4.3.0
regex==2021.4.4
requests==2.22.0
requests-file==1.4.3
requests-ftp==0.3.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
responses==0.13.2
rpm==4.15.1
rsa==3.4.2
ruamel.yaml==0.15.100
s3fs==2021.4.0
s3transfer==0.4.2
scapy==2.4.4
scikit-learn==0.24.2
scipy==1.6.2
seaborn==0.11.1
selenium==3.141.0
selinux==3.0
Send2Trash==1.5.0
sepolicy==3.0
setools==4.3.0
simplegeneric==0.8.1
simpleline==1.6
six==1.14.0
sklearn==0.0
slip==0.6.4
slip.dbus==0.6.4
smartcols==0.3.0
smmap==4.0.0
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcontainers==2.3.0
sos==3.9
soupsieve==1.9.2
Sphinx==3.5.4
sphinx-panels==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.4.11
SSSDConfig==2.3.1
statsmodels==0.12.2
suricata==5.0.3
suricata-update==1.1.2
systemd-python==234
tables==3.6.1
tabulate==0.8.7
Tempita==0.5.1
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
thrift==0.13.0
tldr==0.5
tokenize-rt==4.1.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.61.2
traitlets==4.3.3
typed-ast==1.4.3
typing-extensions==3.7.4.3
urlgrabber==4.1.0
urllib3==1.25.7
userpath==1.4.1
virtualenv==20.4.4
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xarray==0.17.0
xlrd==2.0.1
XlsxWriter==1.4.0
xlwt==1.3.0
xmltodict==0.12.0
yarl==1.6.3
zipp==3.4.1
